pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "active_adaptation"
version = "0.4.2"
authors = [
    {name = "btx0424@SUSTech"}
]
keywords = ["robotics", "rl"]
requires-python = ">=3.11,<3.13"
dependencies = [
    "setproctitle",
    "hydra-core",
    "omegaconf",
    "wandb",
    "moviepy",
    "av",
    "einops",
    "termcolor",
    "pygame",
    "onnxruntime",
    "jaxtyping",
    "tensordict==0.7.0",
    "torchrl==0.7.0",
    "torch==2.7.0",
    "mujoco",
    "linuxfd",
    "pdbpp",
]

[project.scripts]
aa-discover-projects = "active_adaptation.project_loading.discovery:discover_projects"
aa-pull = "active_adaptation.project_loading.cli:aa_pull"

[project.entry-points."active_adaptation.projects"]
facet = "projects.facet"
mimic = "projects.mimic"
metamorph = "projects.metamorph"

[tool.setuptools.packages]
find = { where = ["."], include = [
    "active_adaptation*",
    "scripts*",
    "projects.facet*",
    "projects.mimic*",
    "projects.metamorph*"
    ] }

[tool.setuptools.package-data]
"*" = ["**/*"]

[tool.pyright]
# typeCheckingMode = "strict"
reportIncompatibleMethodOverride = "error"
